phase 8c (part 1/N): decs Archetype.size + iterator widening
First commit on the longarr phase 8c branch. Widens the decs storage
type-level surface to int64 and rewrites the FixedArrayIterator C++
runtime to int64 sizes. This is preparation -- the actual >INT_MAX
runtime scenario is gated by the entityLookup resize cascade
(decsState.entityLookup |> resize(int(id_base) + count) caps total
entities at INT_MAX globally) and EntityId.id : uint. Both follow in
subsequent commits on this same branch before the PR opens.

Changes:

daslib/decs.das
  Archetype.size : int -> int64        (the audit's silent-corruption site)
  entityLookup tuple .index : int -> int64
  get_eid(... index : int) -> int64
  remove_entity(... di : int) -> int64
  decs_array(... capacity : int) -> int64 (routes through _i64 builder)
  entity_count() : int keeps int return + panics if total > INT_MAX
                  (mirrors length() / long_length() surface contract)
  long_entity_count() : int64 NEW -- int64-safe accessor
  for_eid_archetype invoke(blk, ..., int(lookup.index)) cast at boundary
  create_entities_from_cmp invoke(fill_blk, ..., int(eidx), ...) cast at boundary
  get_default_ro / get_optional cap at INT_MAX with panic guard before
                  passing to repeat/repeat_ref (functional.das int64
                  widening is a separate followup)
  Stride multiplications (eidx * c.stride, arch.size * c.stride, etc.)
                  pick up int64(c.stride) casts at every site
  Comparisons with literal 0 widened to 0_l
  range(arch.size) -> range64(0_l, arch.size)

daslib/decs_state.das
  EcsArchetypeView.size : int -> int64
  Switched _builtin_make_temp_array -> _i64 variant for the dump path

daslib/builtin.das
  each(a : auto(TT)[]) passes int64(typeinfo dim(a)) to the widened
  iterator binding -- the only other in-tree caller of
  _builtin_make_fixed_array_iterator besides decs.das.

C++ runtime (the "rewrite said iterator to use 64 bit" piece):
  include/daScript/simulate/runtime_array.h
    FixedArrayIterator::size           : uint32_t -> uint64_t
    SimNode_FixedArrayIterator::size   : uint32_t -> uint64_t
  include/daScript/simulate/aot_builtin.h
    builtin_make_fixed_array_iterator(... int size, ...) -> int64_t size
  src/builtin/module_builtin_runtime.cpp
    Same widening in the impl; negative size clamps to empty (same
    pattern as builtin_make_temp_array_i64).

Tests
  tests/decs/test_archetype.das
    Five comparison literals widened to *_l (existing tests, same logic).
  tests/decs/test_archetype_size_type.das NEW
    static_assert type-contract probes locking Archetype.size : int64,
    long_entity_count() : int64, entity_count() : int. Uses
    stripped_typename so the test doesn't depend on PR #2764 (typeinfo
    is_int64) landing first.
  tests/decs/test_archetype_bulk_create_int64.das NEW
    Small-N (100K + 1K) runtime regression covering bulk-create + delete
    arithmetic neighbors. The genuine >INT_MAX memory-gated probe waits
    for the cascade widening to enable past-INT_MAX creation.

Test status: tests/decs/ 244/244 pass. tests/language/ 1003/1003 pass.
tests/long_array_table/ 45/45 pass.

+++ include/daScript/simulate/aot_builtin.h
@@ -112,7 +112,7 @@ namespace das {
     DAS_API __forceinline bool builtin_iterator_empty ( const Sequence & seq ) { return seq.iter==nullptr; }
 
     DAS_API void builtin_make_good_array_iterator ( Sequence & result, const Array & arr, int stride, Context * context, LineInfoArg * at );
-    DAS_API void builtin_make_fixed_array_iterator ( Sequence & result, void * data, int size, int stride, Context * context, LineInfoArg * at );
+    DAS_API void builtin_make_fixed_array_iterator ( Sequence & result, void * data, int64_t size, int stride, Context * context, LineInfoArg * at );
     DAS_API void builtin_make_range_iterator ( Sequence & result, range rng, Context * context, LineInfoArg * at );
     DAS_API void builtin_make_urange_iterator ( Sequence & result, urange rng, Context * context, LineInfoArg * at );
     DAS_API void builtin_make_range64_iterator ( Sequence & result, range64 rng, Context * context, LineInfoArg * at );
